requirements.txt
openai~=0.28.0
mysql.connector~=0.0.3
chardet~=5.2.0